+++ lib/form/fields/date_picker.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
 import 'package:reactive_forms/reactive_forms.dart';
 import 'package:stoyco_shared/form/forms.dart';
 
@@ -49,14 +50,18 @@ class StoycoDatePicker extends StatelessWidget {
         locale: const Locale('es'),
         builder: (BuildContext context, picker, child) {
           return StoyCoTextFormField(
-              labelText: labelText,
-              hintText: hintText,
-              formControlName: 'birthDate',
-              validationMessages:
-                  validationMessages ?? StoycoForms.validationMessages(),
-              onTap: (value) {
-                picker.showPicker();
-              });
+            labelText: labelText,
+            hintText: hintText,
+            formControlName: 'birthDate',
+            validationMessages:
+                validationMessages ?? StoycoForms.validationMessages(),
+            onTap: (value) {
+              picker.showPicker();
+            },
+            suffixIcon: SvgPicture.asset(
+              'packages/stoyco_shared/lib/assets/icons/calendar.svg',
+            )
+          );
         },
       ),
     );